success-screen: Fixing success_screen__hide_ballot_tracker and success_screen__hide_qr_code options
The activation of the
`election.presentation.extra_options.success_screen__hide_ballot_tracker`
setting should not hide the Download Ballot ticket button in the voting
booth success-screen, but it was doing that so we are fixing it. It also
hiding the QR Code both in the success screen and within the PDF ballot
ticket, so we are fixing that too.