Fix graceful voting period for parent-child elections (#244)
Graceful voting period allows voters that are already logged in
the voting booth to vote even if the election voting period has
ended.

However, this feature didn't work before this bugfix with
parent-children elections, because the voting booth ensured that
the election needs to be within the voting period when retrieving
the election from the backend, and this retrival happens for each
of the children election on demand. When that check did not
checkout, the voter would get an error message stating that the
election is not started.

This commit fixes the issue by allowing the election to be stopped
if the graceful voting period is enabled.