Voting Booth: show ballot warnings in review screen (#322) (#323)
Parent issue: https://github.com/sequentech/epi/issues/111

### Suggestion

When a ballot is invalid implicitly (for example by selecting too many
choices by the voter), the voting screen shows a warning to the voter
telling him that this will make the ballot invalid However, in the
review screen no warning is shown. The proposal is to show this warning
in the review screen too.

### Implementation details and rationale

Note that this feature is already implemented for the [new voting booth
design](https://github.com/sequentech/voting-booth/issues/304) that will
be launched in a couple weeks in version `8.0.0`, but EPI wants to have
access to this feature in version `7.x` to be able to deploy the
breaking changes of version `8.0.0` later with more time and internal
testing.

After a debate seen below, we implemented this feature by saving the
warnings from the ballot selection screen and show them in a similar
manner in the review screen. We will not only show the implicit invalid
vote warning, but more generally the same warnings shown in the ballot
selection screen for consistency.

The change in the review screen can be seen in the screenshot below:


![Image](https://user-images.githubusercontent.com/81968/236503608-708d9515-a691-4298-91c7-485e5770c248.png)

As it can be seen and for visual consistency with respect to the
warnings shown on top of it, we center-aligned the ballot tracker line
on top of the action button in review screen. The warnings are shown in
the bottom section of the screen, that in certain cases (for example if
it's [configured to be anchored to the

bottom](https://sequentech.github.io/documentation/docs/general/reference/election-creation-json#election-presentation-anchor_continue_btn_to_bottom)
or if it's a small screen device) it's part of the anchored and sticky
part of the window, being always shown. That works just like in the
ballot selection screen.

### Tasks

```[tasklist]
# Tasks
- [x] Implement saving the warning from the ballot selection screen and show them in a similar manner in the review screen
```